Move all algorithm files to new 'circuit' submodule:
Change the implementation of the algorithm files so that they now
inherit from a `AbstractAlgorithm` class that defines some common
behaviours of the Algorithm classes.

Remove duplication of code where appropriate after this change.
