+++ qcp/circuits/phase_estimation.py
@@ -1,5 +1,6 @@
 import math
 import cmath
+from qcp.circuits.abstract_circuit import AbstractAlgorithm
 from qcp.matrices import DefaultMatrix, Matrix
 import qcp.gates as g
 from qcp.matrices.types import MATRIX
@@ -102,7 +103,7 @@ def inverse_qft_rotation_gate(size: int, current_qubit: int) -> Matrix:
     return gate
 
 
-class PhaseEstimation:
+class PhaseEstimation(AbstractAlgorithm):
 
     def __init__(self, size: int, unitary: Matrix, eigenvector: Matrix):
         """
@@ -122,9 +123,10 @@ class PhaseEstimation:
         PE.run()
         print(PE.measure())
         """
+        super().__init__(size)
+
         assert is_unitary(unitary), "Matrix must be unitary!"
 
-        self.size = size
         self.unitary = unitary
         self.auxiliary = eigenvector
         self.auxsize = int(math.log2(eigenvector.num_rows))
